Board: activeTiles() method
Method which iterates through board tiles and creates a TreeMap with K,V pair of a number (int) and tile name (string). An intended use of this method is in presenting a list of valid tiles to a player who can then easily select one via the assigned key int. This helps make selections quicker and inputs easier to validate.